remade old arch functional equiv on innerhtml change
